requirements.txt
Flask
latex2sympy
simpy